@@ -1,10 +1,7 @@
 '''
 Background correction module (inherited from PySilCam)
 '''
-
 import numpy as np
-from glob import glob
-from pyopia.pipeline import get_load_function
 
 
 def ini_background(bgfiles, load_function):
@@ -86,19 +83,21 @@ def correct_im_accurate(imbg, imraw):
     highlights if the background or raw images are very poorly obtained
 
     Args:
-      imbg (uint8 or float64)  : background averaged image
-      imraw (uint8 or float64) : raw image
+      imbg (float64)  : background averaged image
+      imraw (float64) : raw image
+      imbg (float64)  : background averaged image
+      imraw (float64) : raw image
 
     Returns:
-      imc (uint8 or float64)   : corrected image, same type as input
+      im_corrected (float64)   : corrected image, same type as input
     '''
 
-    imc = np.float64(imraw) - np.float64(imbg)
-    imc += (255 / 2 - np.percentile(imc, 50))
+    im_corrected = imraw - imbg
+    im_corrected += (1 / 2 - np.percentile(im_corrected, 50))
 
-    imc += 255 - imc.max()
+    im_corrected += 1 - im_corrected.max()
 
-    return imc
+    return im_corrected
 
 
 def correct_im_fast(imbg, imraw):
@@ -110,20 +109,18 @@ def correct_im_fast(imbg, imraw):
     highlights, especially if the background or raw images are not properly obtained
 
     Args:
-      imbg (uint8)  : background averaged image
-      imraw (uint8) : raw image
+      imraw (float64) : raw image
+      imbg (float64)  : background averaged image
 
     Returns:
-      imc (uint8)   : corrected image
+      im_corrected (float64)   : corrected image
     '''
-    imc = imraw - imbg
+    im_corrected = imraw - imbg
 
-    imc += 215
-    imc[imc < 0] = 0
-    imc[imc > 255] = 255
-    imc = np.uint8(imc)
+    im_corrected += 215/255
+    im_corrected = np.clip(im_corrected, 0, 1)
 
-    return imc
+    return im_corrected
 
 
 def shift_and_correct(bgstack, imbg, imraw, stacklength, real_time_stats=False):
@@ -135,162 +132,26 @@ def shift_and_correct(bgstack, imbg, imraw, stacklength, real_time_stats=False):
 
     Args:
         bgstack (list)                  : list of all images in the background stack
-        imbg (uint8)                    : background image
-        imraw (uint8)                   : raw image
+        imbg (float64)                  : background image
+        imraw (float64)                 : raw image
         stacklength (int)               : unsed int here - just there to maintain the same behaviour as
                                           shift_bgstack_fast()
         real_time_stats=False (Bool)    : if True use fast functions, if False use accurate functions
 
     Returns:
         bgstack (list)                  : list of all images in the background stack
-        imbg (uint8)                    : background averaged image
-        imc (uint8)                     : corrected image
+        imbg (float64)                  : background averaged image
+        im_corrected (float64)          : corrected image
     '''
 
     if real_time_stats:
-        imc = correct_im_fast(imbg, imraw)
+        im_corrected = correct_im_fast(imbg, imraw)
         bgstack, imbg = shift_bgstack_fast(bgstack, imbg, imraw, stacklength)
     else:
-        imc = correct_im_accurate(imbg, imraw)
+        im_corrected = correct_im_accurate(imbg, imraw)
         bgstack, imbg = shift_bgstack_accurate(bgstack, imbg, imraw, stacklength)
 
-    return bgstack, imbg, imc
-
-
-def backgrounder(av_window, acquire, bad_lighting_limit=None,
-                 real_time_stats=False):
-    '''
-    Generator which interacts with acquire to return a corrected image
-    given av_window number of frame to use in creating a moving background
-
-    Args:
-        av_window (int)               : number of images to use in creating the background
-        acquire (generator object)    : acquire generator object created by the Acquire class
-        bad_lighting_limit=None (int) : if a number is supplied it is used for throwing away raw images that have a
-                                        standard deviation in colour which exceeds the given value
-
-    Yields:
-        timestamp (timestamp)         : timestamp of when raw image was acquired
-        imc (uint8)                   : corrected image ready for analysis or plotting
-        imraw (uint8)                 : raw image
-
-    Example:
-
-    .. code-block:: python
-
-        avwind = 10 # number of images used for background
-        imgen = backgrounder(avwind,acquire,bad_lighting_limit) # setup generator
-
-        n = 10 # acquire 10 images and correct them with a sliding background:
-        for i in range(n):
-            imc = next(imgen)
-            print(i)
-    '''
-
-    # Set up initial background image stack
-    bgstack, imbg = ini_background(av_window, acquire)
-    stacklength = len(bgstack)
-
-    # Aquire images, apply background correction and yield result
-    for timestamp, imraw in acquire:
-
-        if bad_lighting_limit is not None:
-            bgstack_new, imbg_new, imc = shift_and_correct(bgstack, imbg,
-                                                           imraw, stacklength, real_time_stats)
-
-            # basic check of image quality
-            r = imc[:, :, 0]
-            g = imc[:, :, 1]
-            b = imc[:, :, 2]
-            s = np.std([r, g, b])
-            # ignore bad images
-            if s <= bad_lighting_limit:
-                bgstack = bgstack_new
-                imbg = imbg_new
-                yield timestamp, imc, imraw
-            else:
-                print('bad lighting, std={0}'.format(s))
-        else:
-            bgstack, imbg, imc = shift_and_correct(bgstack, imbg, imraw,
-                                                   stacklength, real_time_stats)
-            yield timestamp, imc, imraw
-
-
-def subtract_background(imbg, imraw):
-    ''' simple background substraction
-
-    Returns
-    -------
-    np.array : image
-        image corrected by simple subtraction (imraw - imbw)
-    '''
-    return imraw - imbg
-
-
-class CreateBackground():
-    '''
-    :class:`pyopia.pipeline` compatible class that calls: :func:`pyopia.background.ini_background`.
-    This runs by default in the pipeline initial steps if named as 'createbackground'.
-
-    Pipeline input data:
-    --------------------
-    :class:`pyopia.pipeline.Data`
-
-        containing the following keys:
-
-        :attr:`pyopia.pipeline.Data.raw_files`
-
-        :attr:`pyopia.pipeline.Data.imc`
-
-        :attr:`pyopia.pipeline.Data.bgstack`
-
-        :attr:`pyopia.pipeline.Data.imraw`
-
-        :attr:`pyopia.pipeline.Data.imbg`
-
-    Parameters:
-    -----------
-    average_window : int
-        number of images to use in the background image stack
-
-    instrument_module: (str, optional)
-        Defaults to 'imread' if not defined
-        Other alternatives are: 'holo' or 'silcam' if you want to use the `load_image`functions
-        implemented within the {mod}`pyopia.instrument` submodule.
-
-    Returns:
-    --------
-    :class:`pyopia.pipeline.Data`
-        containing the following new keys:
-
-        :attr:`pyopia.pipeline.Data.bgstack`
-
-        :attr:`pyopia.pipeline.Data.imbg`
-
-    Example pipeline uses:
-    ----------------------
-
-    .. code-block:: python
-
-        [steps.createbackground]
-        pipeline_class = 'pyopia.background.CreateBackground'
-        average_window = 10
-        instrument_module = 'holo'
-    '''
-
-    def __init__(self, average_window, instrument_module='imread'):
-        self.average_window = average_window
-        self.load_function = get_load_function(instrument_module)
-        pass
-
-    def __call__(self, data):
-        files = glob(data['raw_files'])
-        bgfiles = files[:self.average_window]
-        bgstack, imbg = ini_background(bgfiles, self.load_function)
-
-        data['bgstack'] = bgstack
-        data['imbg'] = imbg
-        return data
+    return bgstack, imbg, im_corrected
 
 
 class CorrectBackgroundAccurate():
@@ -298,17 +159,17 @@ class CorrectBackgroundAccurate():
     :class:`pyopia.pipeline` compatible class that calls: :func:`pyopia.background.correct_im_accurate`
     and will shift the background using a moving average function if given.
 
+    The background stack and background image are created during the first 'average_window' (int) calls
+    to this class, and the skip_next_steps flag is set in the pipeline Data. No background correction
+    is performed during these steps.
+
     Pipeline input data:
     --------------------
     :class:`pyopia.pipeline.Data`
 
         containing the following keys:
 
-        :attr:`pyopia.pipeline.Data.bgstack`
-
         :attr:`pyopia.pipeline.Data.imraw`
-
-        :attr:`pyopia.pipeline.Data.imbg`
 
     Parameters:
     -----------
@@ -320,17 +181,24 @@ class CorrectBackgroundAccurate():
 
         :func:`pyopia.background.shift_bgstack_fast`
 
+    average_window : int
+        number of images to use in the background image stack
+
+    image_source: (str, optional)
+        The key in Pipeline.data of the image to be background corrected.
+        Defaults to 'imraw'
+
     Returns:
     --------
     :class:`pyopia.pipeline.Data`
         containing the following new keys:
 
-        :attr:`pyopia.pipeline.Data.imc`
+        :attr:`pyopia.pipeline.Data.im_corrected`
+        :attr:`pyopia.pipeline.Data.im_corrected`
 
         :attr:`pyopia.pipeline.Data.bgstack`
 
         :attr:`pyopia.pipeline.Data.imbg`
-
 
     Example pipeline uses:
     ----------------------
@@ -341,6 +209,7 @@ class CorrectBackgroundAccurate():
         [steps.correctbackground]
         pipeline_class = 'pyopia.background.CorrectBackgroundAccurate'
         bgshift_function = 'accurate'
+        average_window = 5
 
     Apply static background correction:
 
@@ -349,18 +218,42 @@ class CorrectBackgroundAccurate():
         [steps.correctbackground]
         pipeline_class = 'pyopia.background.CorrectBackgroundAccurate'
         bgshift_function = 'pass'
-
+        average_window = 5
 
     If you do not want to do background correction, leave this step out of the pipeline.
     Then you could use :class:`pyopia.pipeline.CorrectBackgroundNone` if you need to instead.
     '''
 
-    def __init__(self, bgshift_function='pass'):
+    def __init__(self, bgshift_function='pass', average_window=1, image_source='imraw'):
         self.bgshift_function = bgshift_function
-        pass
+        self.average_window = average_window
+        self.image_source = image_source
+
+    def _build_background_step(self, data):
+        '''Add one layer to the background stack from the raw image in data pipeline, and update the background image.'''
+        if 'bgstack' not in data:
+            data['bgstack'] = []
+
+        init_complete = True
+        if len(data['bgstack']) < self.average_window:
+            data['bgstack'].append(data[self.image_source])
+            data['imbg'] = np.mean(data['bgstack'], axis=0)
+            init_complete = False
+
+        return init_complete
 
     def __call__(self, data):
-        data['imc'] = correct_im_accurate(data['imbg'], data['imraw'])
+        # Initialize the background while required bgstack size not reached
+        init_complete = self._build_background_step(data)
+
+        # If we are still building the bgstack, return without doing image correction and bgstack update
+        if not init_complete:
+            # Flag to the pipeline that remaining steps should be skipped since we are still building the background
+            data['skip_next_steps'] = True
+            return data
+
+        data['im_corrected'] = correct_im_accurate(data['imbg'], data[self.image_source])
+        data['im_corrected'] = correct_im_accurate(data['imbg'], data[self.image_source])
 
         match self.bgshift_function:
             case 'pass':
@@ -368,18 +261,19 @@ class CorrectBackgroundAccurate():
             case 'accurate':
                 data['bgstack'], data['imbg'] = shift_bgstack_accurate(data['bgstack'],
                                                                        data['imbg'],
-                                                                       data['imraw'])
+                                                                       data[self.image_source])
             case 'fast':
                 data['bgstack'], data['imbg'] = shift_bgstack_fast(data['bgstack'],
                                                                    data['imbg'],
-                                                                   data['imraw'])
+                                                                   data[self.image_source])
         return data
 
 
 class CorrectBackgroundNone():
     '''
     :class:`pyopia.pipeline` compatible class for use when no background correction is required.
-    This simply makes `data['imc'] = data['imraw'] in the pipeline.
+    This simply makes `data['im_corrected'] = data['imraw'] in the pipeline.
+    This simply makes `data['im_corrected'] = data['imraw'] in the pipeline.
 
     Pipeline input data:
     --------------------
@@ -398,7 +292,7 @@ class CorrectBackgroundNone():
     :class:`pyopia.pipeline.Data`
         containing the following new keys:
 
-        :attr:`pyopia.pipeline.Data.imc`
+        :attr:`pyopia.pipeline.Data.im_corrected`
 
 
     Example pipeline uses:
@@ -416,6 +310,6 @@ class CorrectBackgroundNone():
         pass
 
     def __call__(self, data):
-        data['imc'] = data['imraw']
+        data['im_corrected'] = data['imraw']
 
         return data
