@@ -3,10 +3,25 @@ Module containing tools for classifying particle ROIs
 '''
 
 import os
-
 import numpy as np
 import pandas as pd
-from PIL import Image
+
+import logging
+logger = logging.getLogger()
+
+# import tensorflow here. It must be imported on the processor where it will be used!
+# import is therefore here instead of at the top of file.
+# consider # noqa: E(?) for flake8 / linting
+try:
+    from tensorflow import keras
+    import tensorflow as tf
+except ImportError:
+    info_str = 'ERROR: Could not import Keras. Classify will not work'
+    info_str += ' until you install tensorflow.\n'
+    info_str += 'Use: pip install pyopia[classification]\n'
+    info_str += ' or: pip install pyopia[classification-arm64]'
+    info_str += ' for tensorflow-macos (silicon chips)'
+    raise ImportError(info_str)
 
 
 class Classify():
@@ -55,76 +70,99 @@ class Classify():
         self.model_path = model_path
         self.load_model()
 
+        # Enable this to perform whitebalance correction in the preprocessing step
+        self.correct_whitebalance = False
+
     def __call__(self):
         return self
 
     def load_model(self):
         '''
-        Load the trained tensorflow keras model into the Classify class. example here based on the pysilcam network setup
+        Load a trained Keras model into the Classify class.
 
-        model (tf model object) : loaded tf.keras model from load_model()
+        self.model (tf model object) : loaded Keras model
+        self.class_names (list) : names for the model output classes
         '''
         model_path = self.model_path
-
-        # import tensorflow here. It must be imported on the processor where it will be used!
-        # import is therefore here instead of at the top of file.
-        # consider # noqa: E(?) for flake8 / linting
-        try:
-            from tensorflow import keras
-        except ImportError:
-            info_str = 'ERROR: Could not import Keras. Classify will not work'
-            info_str += ' until you install tensorflow.\n'
-            info_str += 'Use: pip install pyopia[classification]\n'
-            info_str += ' or: pip install pyopia[classification-arm64]'
-            info_str += ' for tensorflow-macos (silicon chips)'
-            raise ImportError(info_str)
 
         os.environ['TF_CPP_MIN_LOG_LEVEL'] = '2'
         keras.backend.clear_session()
 
+        # Instantiate Keras model from file
         path, filename = os.path.split(model_path)
-        header = pd.read_csv(os.path.join(path, 'header.tfl.txt'))
-        self.class_labels = header.columns
         self.model = keras.models.load_model(model_path)
-        print('WARNING: Classification assumes loaded images have values in the range 0-255')
-        return
+
+        # Try to create model output class name list from last model layer name
+        class_labels = None
+        try:
+            class_labels = self.model.layers[-1].name.split('.')
+        except:  # noqa E722
+            logger.info('Could not get class names from model layer name, reverting to old method with header file.')
+
+        # If we could not create correct class names above, revert to old header file method
+        expected_class_number = self.model.layers[-1].output.shape[1]
+        if class_labels is None or len(class_labels) != expected_class_number:
+            header = pd.read_csv(os.path.join(path, 'header.tfl.txt'))
+            class_labels = header.columns
+
+        self.class_labels = class_labels
+        logger.info(self.class_labels)
 
     def preprocessing(self, img_input):
         '''
         Preprocess ROI ready for prediction. example here based on the pysilcam network setup
 
         Args:
-            img_input (uint8)        : a particle ROI before preprocessing
+            img_input (float)        : a particle ROI before preprocessing with range 0-1
 
         Returns:
-            img_preprocessed (uint8) : a particle ROI, corrected and preprocessed, ready for prediction
+            img_preprocessed (float) : a particle ROI with range 0.-255., corrected and preprocessed, ready for prediction
         '''
-        # Scale it to 32x32
-        img_preprocessed = Image.fromarray(img_input)
-        img_preprocessed = img_preprocessed.resize((32, 32), Image.BICUBIC)
-        img_preprocessed = np.array(img_preprocessed)
 
-        # Apply temporary fix for image preprocessing that matches the TFL model conversions in the pysilcam
-        # Tensorflow model that has been adapted for faster use with keras.
-        # Eventual plan is to switch to pytorch and then this can be removed,
-        # but we could not get the same accurancy with pytorch as tensorflow
-        img_preprocessed = (img_preprocessed - 195.17760394934288) / 56.10742134506719
+        whitebalanced = np.copy(img_input).astype(np.float64)
+
+        # Do white-balance correction as a per-channel histogram shift
+        if self.correct_whitebalance:
+            p = 99
+            for c in range(3):
+                whitebalanced[:, :, c] += (p/100) - np.percentile(whitebalanced[:, :, c], p)
+            whitebalanced[whitebalanced > 1] = 1
+            whitebalanced[whitebalanced < 0] = 0
+
+        # convert back to 0-255 scaling (because of this layer in the network:
+        # layers.Rescaling(1./255, input_shape=(img_height, img_width, 3)))
+        # This is useful because it allows training to use tf.keras.utils.image_dataset_from_directory,
+        # which loads images in 0-255 range
+        img = keras.utils.img_to_array(whitebalanced * 255)
+
+        # Get config for image resizing from the model
+        _, img_height, img_width, _ = self.model.get_config()['layers'][0]['config']['batch_shape']
+        pad_to_aspect_ratio = getattr(self.model.layers[0], 'pad_to_aspect_ratio', False)
+
+        # resize to match the dimentions expected by the network
+        img = tf.image.resize(img, [img_height, img_width],
+                              method=tf.image.ResizeMethod.BILINEAR,
+                              preserve_aspect_ratio=pad_to_aspect_ratio)
+
+        img_array = tf.keras.utils.img_to_array(img)
+        img_preprocessed = tf.expand_dims(img_array, 0)  # Create a batch
         return img_preprocessed
 
+    @tf.function
     def predict(self, img_preprocessed):
         '''
         Use tensorflow model to classify particles. example here based on the pysilcam network setup.
 
         Args:
-            img_preprocessed (uint8) : a particle ROI, corrected and preprocessed using :meth:`Classify.preprocessing`,
-            ready for prediction using :meth:`Classify.predict`
+            img_preprocessed (float) : a particle ROI arry, corrected and preprocessed using :meth:`Classify.preprocessing`,
+                                       ready for prediction using :meth:`Classify.predict`
 
         Returns:
             prediction (array)       : the probability of the roi belonging to each class
         '''
 
-        prediction = self.model(np.expand_dims(img_preprocessed, 0))
-
+        prediction = self.model(img_preprocessed, training=False)
+        prediction = tf.nn.softmax(prediction[0])
         return prediction
 
     def proc_predict(self, img_input):
@@ -133,12 +171,11 @@ class Classify():
         using tensorflow model to classify particles. example here based on the pysilcam network setup.
 
         Args:
-            img_input (uint8)  : a particle ROI before preprocessing
+            img_input (float)  : a particle ROI with range 0-1 before preprocessing
 
         Returns:
             prediction (array) : the probability of the roi belonging to each class
         '''
-
         img_preprocessed = self.preprocessing(img_input)
         prediction = self.predict(img_preprocessed)
 
