@@ -68,7 +68,7 @@ class SilCamLoad():
 
     def __call__(self, data):
         timestamp = timestamp_from_filename(data['filename'])
-        img = load_image(data['filename'])
+        img = load_image(data['filename']).astype(np.float64)/255
         data['timestamp'] = timestamp
         data['imraw'] = img
         return data
@@ -89,25 +89,20 @@ class ImagePrep():
     :class:`pyopia.pipeline.Data`
         containing the following new keys:
 
-        :attr:`pyopia.pipeline.Data.imc`
+        :attr:`pyopia.pipeline.Data.im_minimum`
     '''
-    def __init__(self, image_level='imc'):
+    def __init__(self, image_level='im_corrected'):
         self.image_level = image_level
         pass
 
     def __call__(self, data):
         image = data[self.image_level]
-        imc = np.float64(image)
-
-        image = rescale_intensity(image, out_range=(0, 255))
-        data['imref'] = np.uint8(image)
 
         # simplify processing by squeezing the image dimensions into a 2D array
         # min is used for squeezing to represent the highest attenuation of all wavelengths
-        imc = np.min(imc, axis=2)
-        imc /= 255
+        data['im_minimum'] = np.min(image, axis=2)
 
-        data['imc'] = imc
+        data['imref'] = rescale_intensity(image, out_range=(0, 1))
         return data
 
 
@@ -150,10 +145,12 @@ def generate_config(raw_files: str, model_path: str, outfolder: str, output_pref
             },
             'segmentation': {
                 'pipeline_class': 'pyopia.process.Segment',
-                'threshold': 0.85
+                'threshold': 0.85,
+                'segment_source': 'im_minimum'
             },
             'statextract': {
-                'pipeline_class': 'pyopia.process.CalculateStats'
+                'pipeline_class': 'pyopia.process.CalculateStats',
+                'roi_source': 'imref'
             },
             'output': {
                 'pipeline_class': 'pyopia.io.StatsH5',
